Hover in WT working

+++ sw/airborne/modules/delfly_utils/delfly_utils.h
@@ -30,7 +30,7 @@
 
 extern uint8_t LEDS_switch;
 extern uint8_t SRVO_kill;
-extern uint8_t P_WT_heading;
+extern uint16_t P_WT_heading;
 
 
 extern void util_init(void);
